requirements_test.txt
pytest>=7.0.0
pytest-asyncio>=0.18.0
pytest-cov>=3.0.0
pytest-homeassistant-custom-component>=0.12.0